Experimental "Qwerty42" layout, French variant.
This layout is optimized for compact / ergonomic keyboards, where the
position of the `[]` keys could be changed. More generally, this kind of
layout should fit in 42 keys for all printable characters:

- 4 rows of 10 keys, from the number row (1..0) and below (= 40 'sane' keys)
- the 7 (ANSI/pc104) or 8 (ISO/pc105) 'extra' keys must be available in AltGr
- all accented characters should be done with a single dead key (`;:`)
- 40 keys + spacebar + AltGr = 42 keys

The learning curve is a bit steeper and the resulting layout is not as
versatile for other languages (e.g. Esperanto support has been dropped),
but the efficency is better -- especially on my brand new keyboard.io
Model01 (yes, I have one ^_^).

The same principle can be applied to lots of other languages.