better qwerty42-fr + 'Swiss' variant
Turns out that the latest qwerty42-fr layout is not as versatile as I thought, mostly because Italian *does* need an igrave (e.g. `venerdì`).  So let’s have two layouts instead:

- qwerty42-fr is more dedicated to French, though German is also fully supported);
- qwerty42-fr-CH is the more versatile variant: French, Italian, German and Esperanto are fully supported. It could be a good layout for Switzerland. ^_^

The ubreve (`ŭ`), mostly used for Esperanto, has been removed from the qwerty42-fr variant since we lack a dead circumflex for ĉ, ĝ, ĥ, ĵ, ŝ that are used by this language.